// src/models/salesperson.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bespoked Bikes - API
 * BeSpoked Bikes is looking to create a sales tracking application.  BeSpoked is a high-end bicycle shop and each salesperson gets a commission for each bike they sell.  They are introducing a new quarterly bonus based on sales as an incentive.
 *
 * OpenAPI spec version: 1.0.11
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Salesperson
 */
export interface Salesperson {
    /**
     * 
     * @type {number}
     * @memberof Salesperson
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    firstname?: string;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    lastname?: string;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    address?: string;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    begin_date?: string;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    end_date?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Salesperson
     */
    manager?: string;
}